% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_anova_data.R
\docType{data}
\name{anova_data}
\alias{anova_data}
\title{Data sample for the one-way ANOVA example}
\format{
A data frame with 1000 rows, and 2 columns. Suitable for a one-way ANOVA model.
\describe{
  \item{y}{is the continuous dependent variable}
  \item{gr}{is the independent factor (treatments) with four distinct levels}
  }
}
\usage{
data(anova_data)
}
\description{
Data sample for the one-way ANOVA example
}
\examples{
data(anova_data)
}
\author{
V. Vancak
}
\keyword{datasets}
