% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/furukawa.R
\name{furukawa_inc}
\alias{furukawa_inc}
\title{Laupacis' NNT FL type estimator}
\usage{
furukawa_inc(treat, control, tau, d, d.bs, p_c, p_t.boot, p_c.boot, n_c, n_t)
}
\arguments{
\item{treat}{a numeric vector of the treatment arm results}

\item{control}{a numeric vector of the control arm results}

\item{tau}{a scalar that indicates the MCID}

\item{d}{Cohen's delta MLE}

\item{d.bs}{Cohen's delta BS of the MLE estimator}

\item{p_c}{sample proportion of success in the control arm}

\item{p_t.boot}{BS estimator of the sample proportion of success in the treatment arm}

\item{p_c.boot}{BS estimator of the sample proportion of success in the control arm}

\item{n_c}{number of observations in the control arm}

\item{n_t}{number of observations in the treatment arm}

\item{p_t}{sample proportion of success in the treatment arm}
}
\value{
Laupacis' NNT using Furukawa & Leucht's estimator
}
\description{
Internal function. Not for users. Takes two numeric vectors of control and treatment results and returns the Furukawa & Leucht's NNT.
}
