% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laupacis.R
\name{laupacis}
\alias{laupacis}
\title{Non-paramatric Laupacis' NNT}
\usage{
laupacis(treat, control, p_t, p_c, n_t, n_c, p_t.boot, p_c.boot)
}
\arguments{
\item{treat}{a numeric vector of the treatment arm results}

\item{control}{a numeric vector of the control arm results}

\item{p_t}{sample proportion of success in the treatment arm}

\item{p_c}{sample proportion of success in the control arm}

\item{n_t}{number of observations in the treatment arm}

\item{n_c}{number of observations in the control arm}

\item{p_t.boot}{BS estimator of the sample proportion of success in the treatment arm}

\item{p_c.boot}{BS estimator of the sample proportion of success in the control arm}
}
\value{
Laupacis' type NNT non-parametric MLE
}
\description{
Internal function. Not for users. Takes two numeric vectors of control and treatment results
 and returns the non-parametric MLE.
}
