% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_linreg_data.R
\docType{data}
\name{linreg_data}
\alias{linreg_data}
\title{Data sample for the linear regression example}
\format{
A data frame with 1000 rows, and 3 columns. Suitable for a simple linear regression with a dummy variable.
\describe{
  \item{y}{is the continuous dependent variable}
  \item{gr}{is the allocated arm of each subject, where 1 corresponds to treatment, and 0 to control}
  \item{x_var}{is the independent continuous variable}
}
}
\usage{
data(linreg_data)
}
\description{
Data sample for the linear regression example
}
\examples{
data(linreg_data)
}
\author{
V. Vancak
}
\keyword{datasets}
