% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_logreg_data.R
\docType{data}
\name{logreg_data}
\alias{logreg_data}
\title{Data sample for the logistic regression example}
\format{
A data frame with 1000 rows, and 3 columns. Suitable for a logistic regression model.
\describe{
  \item{y}{is the dependent binary 0/1 variable}
  \item{gr}{is the allocated arm of each subject, where 1 corresponds to treatment, and 0 to control}
  \item{x_var}{is the independent continuous variable}
}
}
\usage{
data(logreg_data)
}
\description{
Data sample for the logistic regression example
}
\examples{
data(logreg_data)
}
\author{
V. Vancak
}
\keyword{datasets}
