% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_norm_eq.R
\name{mle_norm_eq_inc}
\alias{mle_norm_eq_inc}
\title{MLE of NNT in equal sd Normal distribution}
\usage{
mle_norm_eq_inc(treat, control, tau, yt_bar, yc_bar, s_ml, p_t.boot, p_c.boot,
  s_ml.bs, n_c, n_t)
}
\arguments{
\item{treat}{a numeric vector of the treatment arm results}

\item{control}{a numeric vector of the control arm results}

\item{tau}{a scalar that indicates the MCID}

\item{yt_bar}{mean value of the treatment arm vector}

\item{yc_bar}{mean value of the control arm vector}

\item{s_ml}{sd of the the pooled variance MLE}

\item{p_t.boot}{BS estimator of the sample proportion of success in the treatment arm}

\item{p_c.boot}{BS estimator of the sample proportion of success in the control arm}

\item{s_ml.bs}{BS estimator of the sd of the the pooled variance MLE}

\item{n_c}{number of observations in the control arm}

\item{n_t}{number of observations in the treatment arm}

\item{p_c}{sample proportion of success in the control arm}

\item{p_t}{sample proportion of success in the treatment arm}
}
\value{
MLE of the Laupacis' type NNT for Normal distribution with equal sd
}
\description{
Internal function. Not for users. Takes two numeric vectors of control and treatment results
 and returns the MLE of the Laupacis' type NNT for Normal distribution with equal variances.
}
