% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_norm_uneq.R
\name{mle_norm_uneq_inc}
\alias{mle_norm_uneq_inc}
\title{MLE of NNT in unequal sd Normal distribution}
\usage{
mle_norm_uneq_inc(treat, control, tau, yt_bar, yc_bar, s_t, s_c, p_t.boot,
  p_c.boot, s_t.bs, s_c.bs, n_c, n_t)
}
\arguments{
\item{treat}{a numeric vector of the treatment arm results}

\item{control}{a numeric vector of the control arm results}

\item{tau}{a scalar that indicates the MCID}

\item{yt_bar}{mean value of the treatment arm vector}

\item{yc_bar}{mean value of the control arm vector}

\item{s_t}{sd of the variance MLE in the treatment arm vector}

\item{s_c}{sd of the variance MLE in the control arm vector}

\item{p_t.boot}{BS estimator of the sample proportion of success in the treatment arm}

\item{p_c.boot}{BS estimator of the sample proportion of success in the control arm}

\item{s_t.bs}{BS estimator of the sd of the variance MLE in the treatment arm vector}

\item{s_c.bs}{BS estimator of the sd of the variance MLE in the control arm vector}

\item{n_c}{number of observations in the control arm}

\item{n_t}{number of observations in the treatment arm}

\item{p_c}{sample proportion of success in the control arm}

\item{p_t}{sample proportion of success in the treatment arm}
}
\value{
MLE of the Laupacis' type NNT for Normal distribution with unequal sd
}
\description{
Internal function. Not for users. Takes two numeric vectors of control and treatment results
 and returns the MLE of the Laupacis' type NNT for Normal distribution with unequal sd.
}
