% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnt_anova.R
\name{nnt_anova}
\alias{nnt_anova}
\title{Laupacis' adjusted and marginal NNT in one-way ANOVA}
\usage{
nnt_anova(response, x, cutoff, base, decrease, data)
}
\arguments{
\item{response}{vector of the response variable (i.e., the dependent variable).}

\item{x}{vector of the explanatory variable, i.e., the factor of different treatments.}

\item{cutoff}{the MCID threshold.}

\item{base}{control group of the x variable in the one-way ANOVA model.}

\item{decrease}{TRUE or FALSE. Indicates whether the MCID change is decrease in the response variable.}

\item{data}{data frame that contains the required variables for the computations.}
}
\value{
The estimated marginal and adjusted NNT with their corresponding 95 percent confidence intervals.
}
\description{
Internal function. Not for users. Calculates Laupacis' adjusted and marginal NNT in one-way ANOVA.
Takes a data-set suitable for a one-way ANOVA analysis and returns
the estimated adjusted NNT(k) and the marginal NNT.
}
\keyword{###}
\keyword{0,}
\keyword{1,}
\keyword{2,}
\keyword{4,}
\keyword{=}
\keyword{DECREASE}
\keyword{FALSE,}
\keyword{INCREASE}
\keyword{SUCCESS}
\keyword{TRUE,}
\keyword{anova_data$gr,}
\keyword{anova_data$y,}
\keyword{anova_data)}
\keyword{base}
\keyword{cutoff}
\keyword{data}
\keyword{data(anova_data)}
\keyword{decrease}
\keyword{internal}
\keyword{nnt_anova(}
\keyword{response}
\keyword{x}
