% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nntkk.R
\name{nnt_kk}
\alias{nnt_kk}
\title{Kraemer & Kupfer's NNT calculator}
\usage{
nnt_kk(type, treat, control, decrease, dist = "none", equal.var = TRUE)
}
\arguments{
\item{type}{specification of the estimation method; 'mle' (for the Maximum Likelihood estimator)
and 'non-param' (for the non-parametric MLE),}

\item{treat}{vector of response variable of the treament group}

\item{control}{vector of response variable of the control group}

\item{decrease}{TRUE or FALSE. Indicates whether the MCID change is decrease in the response variable}

\item{dist}{distribution type (only for the 'mle' type estimator);
"norm" (Normal), "expon" (Exponential). The default value is 'none'}

\item{equal.var}{TRUE or FALSE; Indicates whether the variances are equal - for normal distribution only. The default value is TRUE.}
}
\value{
The estimated Kraemer & Kupfer's NNT and its confidence intervals using the specified estimation method.
}
\description{
Calculates the parametric and the non-parametric KK-NNT
Takes two numeric vectors, treatment and control, and returns
the estimated KK-NNT using the specified method.
}
\examples{
nnt_kk( type     = "non-param",
treat     = rnorm(100, 100, 10),
control   = rnorm(100, 110, 20),
decrease  = TRUE )
}
