% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnt_lm.R
\name{nnt_lm}
\alias{nnt_lm}
\title{Laupacis' adjusted and marginal NNT in linear regression model}
\usage{
nnt_lm(response, x, cutoff, group, decrease, adj, data)
}
\arguments{
\item{response}{vector of the response variable (i.e., the dependent variable).}

\item{x}{vector of the explanatory variable.}

\item{cutoff}{the MCID threshold.}

\item{group}{allocated arm variable where 1 corresponds to the treatment arm, and 0 to the control arm.}

\item{decrease}{TRUE or FALSE. Indicates whether the MCID change is decrease in the response variable}

\item{adj}{value that the NNT need to be adjusted for. The default value is the mean of x.}

\item{data}{data frame that contains the required variables for the computations.}
}
\value{
The estimated marginal and adjusted NNT with their corresponding 95 percent confidence intervals.

data(linreg_data)
}
\description{
Internal function. Not for users. Calculates Laupacis' adjusted and marginal NNT
in linear regression model.
Takes a data-set suitable for a simple linear regression with a dummy variable and an
interaction term, and returns the estimated adjusted NNT(x) and the marginal NNT.
}
\keyword{internal}
