% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnt_logreg.R
\name{nnt_logreg}
\alias{nnt_logreg}
\title{Laupacis' adjusted and marginal NNT in logistic regression model}
\usage{
nnt_logreg(response, x, group, adj, data)
}
\arguments{
\item{response}{vector of the response variable (i.e., the dependent variable).}

\item{x}{vector of the explanatory variable.}

\item{group}{allocated arm variable where 1 corresponds to the treatment arm, and 0 to the control arm.}

\item{adj}{value that the NNT need to be adjusted for. The default value is the mean of x.}

\item{data}{data frame that contains the required variables for the computations.}
}
\value{
The estimated marginal and adjusted NNT with their corresponding 95 percent confidence intervals.
}
\description{
Internal function. Not for users. Calculates Laupacis' adjusted and marginal NNT
in logistic regression model.
Takes a data-set suitable for a logistic regression with a dummy variable and an
interaction term, and returns the estimated adjusted NNT(x) and the marginal NNT.
}
\keyword{)}
\keyword{2,}
\keyword{=}
\keyword{adj}
\keyword{data}
\keyword{group}
\keyword{internal}
\keyword{load("logreg_data.RData"}
\keyword{logreg_data}
\keyword{logreg_data$gr,}
\keyword{logreg_data$x_var,}
\keyword{logreg_data$y,}
\keyword{nnt_logreg(}
\keyword{response}
\keyword{x}
