% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nntkk_exp.R
\name{nntkk_exp_inc}
\alias{nntkk_exp_inc}
\title{MLE of KK-NNT in Exponential distribution}
\usage{
nntkk_exp_inc(treat, control, yt_bar, yc_bar, p_t.boot, p_c.boot, n_c, n_t)
}
\arguments{
\item{treat}{a numeric vector of the treatment arm results}

\item{control}{a numeric vector of the control arm results}

\item{yt_bar}{mean value of the treatment arm vector}

\item{yc_bar}{mean value of the control arm vector}

\item{p_t.boot}{BS estimator of the sample proportion of success in the treatment arm}

\item{p_c.boot}{BS estimator of the sample proportion of success in the control arm}

\item{n_c}{number of observations in the control arm}

\item{n_t}{number of observations in the treatment arm}
}
\value{
MLE of the KK-NNT for the Exponential distribution
}
\description{
Internal function. Not for users. Takes two numeric vectors of control and treatment results
 and returns the MLE of the Kraemer & Kupfer's type NNT for Exponential distribution.
}
