% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nntkk_nonparam.R
\name{nntkk_nonparam}
\alias{nntkk_nonparam}
\title{Non-Parametric MLE of KK-NNT in Exponential distribution}
\usage{
nntkk_nonparam(treat, control, p_c.boot, p_tk)
}
\arguments{
\item{treat}{a numeric vector of the treatment arm results}

\item{control}{a numeric vector of the control arm results}

\item{p_c.boot}{BS estimator of the sample proportion of success in the control arm}

\item{p_tk}{sample estimator of P(Y_t > Y_c)}
}
\value{
Non-Paramaetric MLE of the KK-NNT
}
\description{
Internal function. Not for users. Takes two numeric vectors of control and treatment results
 and returns the non-parametric MLE of the Kraemer & Kupfer's NNT.
}
