% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nntkk_norm_eq.R
\name{nntkk_norm_eq_inc}
\alias{nntkk_norm_eq_inc}
\title{MLE of the KK-NNT in Normal distribution with equal variances}
\usage{
nntkk_norm_eq_inc(treat, control, d, s_ml, p_t.boot, p_c.boot, s_ml.bs, n_c,
  n_t)
}
\arguments{
\item{treat}{a numeric vector of the treatment arm results}

\item{control}{a numeric vector of the control arm results}

\item{d}{Cohen's delta MLE}

\item{s_ml}{pooled sd MLE}

\item{p_t.boot}{BS estimator of the sample proportion of success in the treatment arm}

\item{p_c.boot}{BS estimator of the sample proportion of success in the control arm}

\item{s_ml.bs}{BS of the pooled sd MLE}

\item{n_c}{number of observations in the control arm}

\item{n_t}{number of observations in the treatment arm}
}
\value{
MLE of the KK-NNT for the Normal distribution with equal sd
}
\description{
Internal function. Not for users. Takes two numeric vectors of control and treatment results
 and returns the MLE of the Kraemer & Kupfer's type NNT for Normal distribution with equal sd.
}
