% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nntkk_norm_uneq.R
\name{nntkk_norm_uneq_inc}
\alias{nntkk_norm_uneq_inc}
\title{MLE of the KK-NNT in Normal distribution with unequal sd}
\usage{
nntkk_norm_uneq_inc(yt_bar, yc_bar, s_t, s_c, p_t.boot, p_c.boot, n_c, n_t)
}
\arguments{
\item{yt_bar}{mean value of the treatment arm vector}

\item{yc_bar}{mean value of the control arm vector}

\item{s_t}{sd of the treatment arm vector}

\item{s_c}{sd of the control arm vector}

\item{p_t.boot}{BS estimator of the sample proportion of success in the treatment arm}

\item{p_c.boot}{BS estimator of the sample proportion of success in the control arm}

\item{n_c}{number of observations in the control arm}

\item{n_t}{number of observations in the treatment arm}
}
\value{
MLE of the KK-NNT for the Normal distribution with unequal sd
}
\description{
Internal function. Not for users. Takes two numeric vectors of control and treatment results
 and returns the MLE of the Kraemer & Kupfer's type NNT for Normal distribution with unequal sd.
}
