% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nntl.R
\name{nnt_l}
\alias{nnt_l}
\title{Laupacis' unadjusted NNT calculator}
\usage{
nnt_l(type, treat, control, cutoff, decrease, dist = "none", equal.var = TRUE)
}
\arguments{
\item{type}{specification of the estimation method; 'mle' (for the Maximum Likelihood estimator),
'fl' (for Furukawa & Leucht's estimator),
'laupacis' (for the non-parametric MLE estimator)}

\item{treat}{vector of response variable of the treatment group}

\item{control}{vector of response variable of the control group}

\item{cutoff}{a scalar that is the MCID}

\item{decrease}{TRUE or FALSE. Indicates whether the MCID change is decrease in the response variable}

\item{dist}{distribution type (if specified); "normal" (Normal), "expon" (Exponential).
The default value is 'none'.}

\item{equal.var}{TRUE or FALSE; Indicates whether the variances are equal - for normal distribution only.
The default value is TRUE.}
}
\value{
The estimated Laupacis' NNT and its confidence intervals using the specified estimation method.
}
\description{
Calculates Laupacis' type NNT.
Takes two numeric vectors that are the outcomes of the treatment and control arms,
and returns the estimated NNT using the specified estimation method.
}
\examples{
 nnt_l( type      = "mle",
        treat     = rnorm(1000, 110, 10),
        control   = rnorm(1000, 100, 10),
        cutoff    = 100,
        equal.var = TRUE,
        dist      = "normal",
        decrease  = FALSE )

}
\references{
Laupacis, A., Sackett, D. L., & Roberts, R. S. (1988). An assessment of clinically useful measures of the consequences of treatment. New England journal of medicine, 318(26), 1728-1733.

Vancak, V., Goldberg, Y., & Levine, S. Z. (2020). Systematic analysis of the number needed to treat. Statistical Methods in Medical Research, 0962280219890635.

Furukawa, T. A., & Leucht, S. (2011). How to obtain NNT from Cohen's d: comparison of two methods. PloS one, 6(4), e19070.
}
