% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nnt_lm.R
\name{nnt_lm}
\alias{nnt_lm}
\title{Laupacis' adjusted and marginal NNT in linear regression model}
\usage{
nnt_lm(response, x, cutoff, group, decrease, adj, data)
}
\arguments{
\item{response}{vector of the response variable (i.e., the dependent variable).}

\item{x}{vector of the explanatory variable.}

\item{cutoff}{the MCID threshold.}

\item{group}{allocated arm variable where 1 corresponds to the treatment arm, and 0 to the control arm.}

\item{decrease}{TRUE or FALSE. Indicates whether the MCID change is decrease in the response variable}

\item{adj}{value that the NNT need to be adjusted for. The default value is the mean of x.}

\item{data}{data frame that contains the required variables for the computations.}
}
\value{
The estimated marginal and adjusted NNT with their corresponding 95 percent confidence intervals.

data(linreg_data)

nnt_lm( response = linreg_data$y,
        x        = linreg_data$x_var,
        cutoff   = 3,
        group    = linreg_data$gr,
        decrease = F,
        adj      = 2.6,
        data     = linreg_data )

inv_dat = data.frame( y = linreg_data$y, x_var = linreg_data$x_var, gr = 1 - linreg_data$gr )

nnt_lm( response = inv_dat$y,
        x        = inv_dat$x_var,
        cutoff   = 3,
        group    = inv_dat$gr,
        decrease = T,
        adj      = 2.6,
        data     = inv_dat )
}
\description{
Internal function. Not for users. Calculates Laupacis' adjusted and marginal NNT
in linear regression model.
Takes a data-set suitable for a simple linear regression with a dummy variable and an
interaction term, and returns the estimated adjusted NNT(x) and the marginal NNT.
}
\keyword{internal}
